fix(api): stop truncating persons.age to int before or None
persons.age already stores CampMinder's yy.mm as a REAL (7 years 4
months is literally 7.04). int() truncated the fractional months off
before `or None` ever saw the value, so infants at 0.01-0.11 years
were truncated to 0 and rendered blank -- the same collapse that is
correct and deliberate for the genuinely unknown-age population (raw
age exactly 0.0, no birthdate on file).

Remove the int() wrapper, keep the or None: on the raw float,
0.06 or None -> 0.06 (an infant keeps its age) and 0.0 or None ->
None (unknown age stays unknown). Also fix the oldest-first sort key
to compare the float instead of the truncated int, so e.g. 1.11
correctly sorts above 1.02.

PartyChild.age moves from int | None to float | None in the response
schema to carry the fractional value through -- Pydantic's strict int
coercion rejects a fractional float outright, so this is required for
the fix to function, not an unrelated change. The generated frontend
type stays `number | null` either way, so no api-types regen is
needed.

Closes #2088.

+++ api/schemas/lodging.py
@@ -171,7 +171,7 @@ class PartyChild(BaseModel):
 
     person_cm_id: int = 0
     display_name: str = ""
-    age: int | None = None
+    age: float | None = None
     grade: int | None = None
